feat(userService)：用户信息修改，用户权限管理

+++ backend/UkraineAid/Service/SystemService/src/main/java/com/byb/systemservice/Entity/Syslog.java
@@ -30,7 +30,7 @@ public class Syslog {
     private Long objId;
 
     @TableField("OPERATOR")
-    private int operator;
+    private Long operator;
 
     @TableField("MESSAGE")
     private String message;
