+++ backend/UkraineAid/Service/SystemService/src/main/java/com/byb/systemservice/Vo/SyslogForm.java
@@ -16,7 +16,7 @@ public class SyslogForm {
 
     private Long objId;
 
-    private int operator;
+    private Long operator;
 
     private String message;
